chore: use different files for store trait impl
