--- crates/proksi/src/stores/challenges.rs
@@ -1,5 +1,0 @@
-// DEPRECATED: This is kept for backward compatibility but is no longer used directly.
-// Challenge storage is now handled through the Store trait in adapter.rs
-
-// Type definition for the challenge store (token, proof)
-// pub type ChallengeStore = papaya::HashMap<String, (String, String)>;
